+++ Connectors/Reqnroll.VisualStudio.ReqnrollConnector.Generic/Discovery/IBindingProvider.cs
@@ -1,0 +1,11 @@
+﻿using System.Reflection;
+using Reqnroll.Bindings.Provider.Data;
+using ReqnrollConnector.AssemblyLoading;
+using ReqnrollConnector.Logging;
+
+namespace ReqnrollConnector.Discovery;
+
+public interface IBindingProvider
+{
+    BindingData DiscoverBindings(TestAssemblyLoadContext testAssemblyContext, Assembly testAssembly, string? configFileContent, ILogger log);
+}