+++ Connectors/Reqnroll.VisualStudio.ReqnrollConnector.Generic/Logging/IAnalyticsContainer.cs
@@ -3,5 +3,5 @@ namespace ReqnrollConnector.Logging;
 public interface IAnalyticsContainer
 {
     void AddAnalyticsProperty(string key, string value);
-    ImmutableSortedDictionary<string, string> ToImmutable();
+    Dictionary<string, object> ToDictionary();
 }