+++ Connectors/Reqnroll.VisualStudio.ReqnrollConnector.Generic/ReqnrollProxies/IBindingRegistryAdapter.cs
@@ -1,7 +1,0 @@
-namespace ReqnrollConnector.ReqnrollProxies;
-
-public interface IBindingRegistryAdapter
-{
-    IEnumerable<StepDefinitionBindingAdapter> GetStepDefinitions();
-    IEnumerable<HookBindingAdapter> GetHooks();
-}