+++ Connectors/Reqnroll.VisualStudio.ReqnrollConnector.Generic/ReqnrollProxies/IBindingRegistryFactory.cs
@@ -1,9 +1,0 @@
-namespace ReqnrollConnector.ReqnrollProxies;
-
-public interface IBindingRegistryFactory
-{
-    IBindingRegistryAdapter GetBindingRegistry(
-        AssemblyLoadContext assemblyLoadContext,
-        Assembly testAssembly,
-        Option<FileDetails> configFile);
-}